--- src/FeedState/FeedState.interface.ts
@@ -1,13 +1,0 @@
-import { RssItem } from "../parser";
-
-export interface Feed {
-    url: string;
-    lastFetched: Date;
-    items: RssItem[];
-    viewedItems?: RssItem[];
-}
-
-export interface FeedState {
-    feeds: Feed[];
-    activeFeed: Feed | null;
-}
