+++ tests/FeedStateTest/FeedStateTest.ts
@@ -1,8 +1,8 @@
 // test/FeedState.test.ts
 
 import { expect, test } from "@playwright/test";
-import { addFeed, setActiveFeed, markItemAsViewed, getFeedState } from "../../src/FeedState/feedState";
-import { fetchAndParseRss } from "../../src/parser";
+import { addFeed, setActiveFeed, markItemAsViewed, getFeedState } from "../../src/State/State";
+import { fetchAndParseRss } from "../../src/Parser/fetchAndParseRss";
 
 const feedUrl = "https://feeds.npr.org/1001/rss.xml"; // You can change this URL to test with different feeds
 const nonExistentFeedUrl = feedUrl.replace(".org", ".com"); // This should be a URL that doesn't exist in the feeds
